Manufacturer, Component, Feature, Feature Type, Componenet Feature Type, Item, Item Feature almost done

+++ src/main/java/com/primebuild_online/model/DTO/ItemReqDTO.java
@@ -15,7 +15,7 @@ public class ItemReqDTO {
     private String itemName;
     private Integer quantity;
     private Double price;
-    private Integer componentId;
-    private Integer manufacturerId;
+    private Long componentId;
+    private Long manufacturerId;
     private List<Feature> featureList = new ArrayList<>();
 }
